+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [UNRELEASED]
+### Added
+- Includes the number of the Dependabot pull request at the end of each entry line. A sample line is below
+  - `- Bump <package> from <old-version> to <new-version> (#pr)`
 
 ## [3.0.0]
 ### Added